+++ jgclark.Summaries/CHANGELOG.md
@@ -2,8 +2,7 @@
 (And see the full [README](https://github.com/NotePlan/plugins/tree/main/jgclark.Summaries).)
 ## [0.11.0] - 2022-06-27 (@@@ unreleased)
 ### Changed
-- speeded up the **/saveSearchResults** command. (Under the hood the plugin now uses an API that takes advantage of caching.)
-- now trims the display of matching results in search output, but still highlights the matched terms
+- moved the **/saveSearchResults** **/saveSearchResultsInPeriod** commands to a separate **SearchHelpers** plugin.
 
 ## [0.10.0] - 2022-06-26
 ### Changed